[DDW-1111] Use 0 as default value for asset decimal user setting
