+++ source/renderer/app/api/api.ts
@@ -223,11 +223,11 @@ import type {
   ApiAsset,
   StoredAssetMetadata,
 } from './assets/types';
-import type { AssetLocalData } from './utils/localStorage';
 import Asset from '../domains/Asset';
 import { getAssets } from './assets/requests/getAssets';
 import { getAccountPublicKey } from './wallets/requests/getAccountPublicKey';
 import { doesWalletRequireAdaToRemainToSupportTokens } from './utils/apiHelpers';
+import { AssetLocalData, defaultAssetLocalData } from '../types/localDataTypes';
 
 export default class AdaApi {
   config: RequestConfig;
@@ -670,7 +670,8 @@ export default class AdaApi {
       const assets = response.map((asset) =>
         _createAssetFromServerData(
           asset,
-          assetsLocaldata[asset.policy_id + asset.asset_name] || {},
+          assetsLocaldata[asset.policy_id + asset.asset_name] ||
+            defaultAssetLocalData,
           this.storedAssetMetadata
         )
       );
