+++ source/renderer/app/api/utils/localStorage.ts
@@ -1,33 +1,34 @@
 /* eslint-disable consistent-return */
-import { includes, without, get } from 'lodash';
+import { get, includes, without } from 'lodash';
 import { toJS } from '../../../../common/utils/helper';
 import { electronStoreConversation } from '../../ipc/electronStoreConversation';
+import type { WalletMigrationStatus } from '../../stores/WalletMigrationStore';
 import { WalletMigrationStatuses } from '../../stores/WalletMigrationStore';
 import {
-  STORAGE_TYPES as types,
   STORAGE_KEYS as keys,
+  STORAGE_TYPES as types,
 } from '../../../../common/config/electron-store.config';
 import type { NewsTimestamp } from '../news/types';
-import type { WalletMigrationStatus } from '../../stores/WalletMigrationStore';
 import type {
-  TransportDevice,
   HardwareWalletExtendedPublicKeyResponse,
-  DeviceType,
+  TransportDevice,
 } from '../../../../common/types/hardware-wallets.types';
 import type { StorageKey } from '../../../../common/types/electron-store.types';
 import type { Currency, DeprecatedCurrency } from '../../types/currencyTypes';
 import {
-  CURRENCY_IS_ACTIVE_BY_DEFAULT,
   CURRENCY_DEFAULT_SELECTED,
+  CURRENCY_IS_ACTIVE_BY_DEFAULT,
 } from '../../config/currencyConfig';
+import {
+  AssetLocalData,
+  HardwareWalletDevicesType,
+  HardwareWalletLocalData,
+  HardwareWalletsLocalData,
+  UnpairedHardwareWalletData,
+  WalletLocalData,
+  defaultAssetLocalData,
+} from '../../types/localDataTypes';
 
-export type WalletLocalData = {
-  id: string;
-  recoveryPhraseVerificationDate?: Date | null | undefined;
-  creationDate: Date;
-  showUsedAddresses: boolean;
-};
-export type WalletsLocalData = Record<StorageKey, WalletLocalData>;
 export type SetHardwareWalletLocalDataRequestType = {
   walletId: string;
   data: {
@@ -42,27 +43,6 @@ export type SetHardwareWalletDeviceRequestType = {
   data: UnpairedHardwareWalletData;
 };
 
-export type UnpairedHardwareWalletData = {
-  deviceType?: DeviceType;
-  deviceModel?: string;
-  deviceName?: string;
-  path?: string | null | undefined;
-  paired?: string | null | undefined;
-  disconnected?: boolean;
-};
-
-export type HardwareWalletLocalData = {
-  id: string;
-  deviceType: DeviceType;
-  device: TransportDevice;
-  extendedPublicKey: HardwareWalletExtendedPublicKeyResponse;
-  disconnected: boolean;
-};
-export type HardwareWalletsLocalData = Record<string, HardwareWalletLocalData>;
-export type HardwareWalletDevicesType = Record<string, TransportDevice>;
-export type AssetLocalData = {
-  decimals: number;
-};
 /**
  * This api layer provides access to the electron local storage
  * for user settings that are not synced with any coin backend.
@@ -298,7 +278,11 @@ export default class LocalStorageApi {
     policyId: string,
     assetName: string
   ): Promise<AssetLocalData> =>
-    LocalStorageApi.get(keys.ASSET_DATA, {}, `${policyId}${assetName}`);
+    LocalStorageApi.get(
+      keys.ASSET_DATA,
+      defaultAssetLocalData,
+      `${policyId}${assetName}`
+    );
   setAssetLocalData = (
     policyId: string,
     assetName: string,
