+++ source/renderer/app/stores/HardwareWalletsStore.ts
@@ -84,12 +84,8 @@ import type {
   VotingDataType,
 } from '../api/transactions/types';
 import type {
-  HardwareWalletLocalData,
-  HardwareWalletsLocalData,
-  HardwareWalletDevicesType,
   SetHardwareWalletLocalDataRequestType,
   SetHardwareWalletDeviceRequestType,
-  UnpairedHardwareWalletData,
 } from '../api/utils/localStorage';
 import type {
   TransportDevice,
@@ -101,6 +97,12 @@ import type {
   TrezorWitness,
 } from '../../../common/types/hardware-wallets.types';
 import { logger } from '../utils/logging';
+import {
+  HardwareWalletDevicesType,
+  HardwareWalletLocalData,
+  HardwareWalletsLocalData,
+  UnpairedHardwareWalletData,
+} from '../types/localDataTypes';
 
 export type TxSignRequestTypes = {
   coinSelection: CoinSelectionsResponse;
